Moved code to destination list view
