+++ SwiftData_iTour/DestinationListingView.swift
@@ -6,11 +6,38 @@
 //
 
 import SwiftUI
+import SwiftData
 
 struct DestinationListingView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+  @Environment(\.modelContext) private var modelContext
+  @Query(sort: [SortDescriptor(\Destination.priority, order: .reverse),
+                SortDescriptor(\Destination.name)],
+         animation: .bouncy) private var destinations: [Destination]
+   
+  var body: some View {
+    List {
+      ForEach(destinations) { destination in
+        NavigationLink(value: destination) {
+          VStack(alignment: .leading) {
+            Text(destination.name)
+              .font(.headline)
+            Text(destination.date.formatted(
+              date: .long,
+              time: .shortened
+            ))
+          }
+        }
+      }
+      .onDelete(perform: deleteDestinations(_:))
     }
+  }
+  
+  private func deleteDestinations(_ indexSet: IndexSet) {
+    for index in indexSet {
+      let destination = destinations[index]
+      modelContext.delete(destination)
+    }
+  }
 }
 
 #Preview {
